finalized backend Clinica

+++ back-end/src/controllers/MedicController.js
@@ -1,19 +1,28 @@
-const Medic = require('../models/Medic')
+const Medic = require('../models/Medic');
+const Owner = require('../models/Owner');
 
 const MedicController = {
 
-    async create (req, res) {
-        try {
-            const create = await Medic.create(req.body);
-            return res.json({ Status: "Create Sucess", create }).end()
+    async create(req, res) {
 
+        const { auth } = req.body;
+        var _id = auth;
+        const acesso = await Owner.findOne({ _id })
+
+        if (acesso) {
+            const create = await Medic.create(req.body);
+            if (create)
+                return res.json(create.name + "created")
         }
-        catch (err) {
-            return res.json().end();
+        else {
+            return res.json({ Status: "Authorization recused" });
         }
+
+        return res.status(500).json({ error: 'Error' }).end();
+
     },
 
-    async read (req, res) {
+    async read(req, res) {
         if (Object.keys(req.query).length > 0) {
             this.read(req, res).end()
         } else {
@@ -26,7 +35,7 @@ const MedicController = {
         }
     },
 
-    async getOne  (req, res) {
+    async getOne(req, res) {
         // if(Object.keys(req.query).length > 0){
         //     this.read(req, res)
 
@@ -45,21 +54,25 @@ const MedicController = {
         }
     },
 
-    async update (req, res) {
-        // const id = req.params.id
-        // const user = await Medic.findByIdAndUpdate(id)
-
-        // return res.send(user)
+    async update(req, res) {
 
         try {
-            const id = req.body._id
-            const obj = await Medic.findByIdAndUpdate(id, req.body)
-            if (obj) {// obj foi encontrado
-                //HTTP 204: No content
+            const { auth } = req.body;
+            var _id = auth;
+            const acesso = await Owner.findOne({ _id })
 
-                return res.send(obj).end()
-            } else {
-                return res.send(obj).end()
+            if (acesso) {
+                const id = req.body._id
+                const obj = await Medic.findByIdAndUpdate(id, req.body)
+                if (obj) {
+
+                    return res.send(obj).end()
+                } else {
+                    return res.send(obj).end()
+                }
+            }
+            else{
+                res.json({Status: "Ascess recused"})
             }
         }
         catch (erro) {
@@ -68,22 +81,31 @@ const MedicController = {
         }
     },
 
-    async remove (req, res) {
+    async remove(req, res) {
         try {
-            const id = req.body._id
-            const obj = await Medic.findByIdAndDelete(id)
-            if (obj) {
-                res.json({ status: `${obj.name} removed` })
-            } else {
-                res.json({ status: `${obj.id} not found` })
+            const { auth } = req.body;
+            var _id = auth;
+            const acesso = await Owner.findOne({ _id })
+
+            if (acesso) {
+                const id = req.body._id
+                const obj = await Medic.findByIdAndDelete(id)
+                if (obj) {
+                    res.json({ status: `${obj.name} removed` })
+                } else {
+                    res.json({ status: `${obj.id} not found` })
+                }
+            }
+            else {
+                res.json({ Status: "Ascess recused" })
             }
         }
         catch (err) {
-            res.json({ status: `error` })
+            res.json({ status: `error`, err })
 
         }
     }
 }
-            
+
 
 module.exports = MedicController;
